Revert "Merge pull request #151 from blehnen/sqlite-outbox-149"
This reverts commit 33aa6e99533c4538640ba44d0cec1fd0ff1a24ac, reversing
changes made to 26a405af07bd4e0f1343aa9c0a3a66906b635ce3.
