+++ Source/DotNetWorkQueue.Transport.SQLite/Basic/CommandHandler/SendMessageCommandHandler.cs
@@ -18,8 +18,6 @@
 // ---------------------------------------------------------------------
 using System;
 using System.Data;
-using System.Data.SQLite;
-using System.Diagnostics.CodeAnalysis;
 using DotNetWorkQueue.Configuration;
 using DotNetWorkQueue.Exceptions;
 using DotNetWorkQueue.Transport.RelationalDatabase.Basic;
@@ -122,9 +120,6 @@ namespace DotNetWorkQueue.Transport.SQLite.Basic.CommandHandler
                 _messageExpirationEnabled = _options.Value.EnableMessageExpiration;
             }
 
-            if (commandSend is RelationalSendMessageCommand relCommand && relCommand.ExternalTransaction != null)
-                return HandleExternalTransaction(commandSend);
-
             using (var connection = _dbFactory.CreateConnection(_configurationSend.ConnectionInfo.ConnectionString, false))
             {
                 connection.Open();
@@ -223,117 +218,6 @@ namespace DotNetWorkQueue.Transport.SQLite.Basic.CommandHandler
                     return id;
                 }
             }
-        }
-
-        /// <summary>
-        /// Caller-supplied-transaction fork of <see cref="Handle(SendMessageCommand)"/>. Reuses
-        /// the caller's transaction and its connection for all queue INSERTs; never commits,
-        /// rolls back, closes, or disposes the caller's resources. Invoked from
-        /// <see cref="Handle"/> when <see cref="RelationalSendMessageCommand.ExternalTransaction"/>
-        /// is non-null. The producer surface (<c>SqliteRelationalProducerQueue&lt;T&gt;</c>)
-        /// guarantees the transaction is a <c>System.Data.SQLite.SQLiteTransaction</c> and its
-        /// connection's database matches the queue's configured database via the validator at
-        /// the API boundary, so this method performs no validation of its own.
-        /// </summary>
-        /// <remarks>
-        /// The handler operates on <see cref="IDbTransaction"/> / <see cref="IDbConnection"/>
-        /// abstractions only — no downcast to the sealed <c>System.Data.SQLite</c> provider
-        /// types. Same discipline applied across the PG/SS transports (CLAUDE.md sealed-cast
-        /// lesson) so the method remains unit-testable with NSubstitute on the interfaces.
-        /// </remarks>
-        /// <param name="commandSend">The send-message command carrying a non-null
-        /// <see cref="RelationalSendMessageCommand.ExternalTransaction"/>.</param>
-        /// <returns>The newly-inserted message ID.</returns>
-        /// <exception cref="DotNetWorkQueueException">Thrown when the INSERT returns a zero ID
-        /// or when the job-uniqueness query rejects the command.</exception>
-        [SuppressMessage("Microsoft.Security", "CA2100:Review SQL queries for security vulnerabilities", Justification = "Query OK")]
-        private long HandleExternalTransaction(SendMessageCommand commandSend)
-        {
-            // Cast guard: Handle() only routes here for RelationalSendMessageCommand
-            // (the `commandSend is RelationalSendMessageCommand` pattern in the fork check),
-            // so this cast is always safe. The producer subclass also enforces the
-            // SQLiteTransaction subtype via GuardSQLiteTransaction before construction.
-            var relCommand = (RelationalSendMessageCommand)commandSend;
-            IDbTransaction transaction = relCommand.ExternalTransaction;
-            IDbConnection connection = transaction.Connection;
-
-            var expiration = TimeSpan.Zero;
-            if (_messageExpirationEnabled.Value)
-            {
-                expiration = MessageExpiration.GetExpiration(commandSend, data => data.GetExpiration());
-            }
-
-            var jobName = _jobSchedulerMetaData.GetJobName(commandSend.MessageData);
-            var scheduledTime = DateTimeOffset.MinValue;
-            var eventTime = DateTimeOffset.MinValue;
-            if (!string.IsNullOrWhiteSpace(jobName))
-            {
-                scheduledTime = _jobSchedulerMetaData.GetScheduledTime(commandSend.MessageData);
-                eventTime = _jobSchedulerMetaData.GetEventTime(commandSend.MessageData);
-            }
-
-            if (!(string.IsNullOrWhiteSpace(jobName) ||
-                  _jobExistsHandler.Handle(new DoesJobExistQuery<IDbConnection, IDbTransaction>(
-                      jobName, scheduledTime, connection, transaction)) == QueueStatuses.NotQueued))
-            {
-                throw new DotNetWorkQueueException(
-                    "Failed to insert record - the job has already been queued or processed");
-            }
-
-            long id;
-            using (var command = SendMessage.GetMainCommand(commandSend, connection, _commandCache, _headers, _serializer))
-            {
-                command.Transaction = transaction;
-                command.ExecuteNonQuery();
-                using (var commandId = connection.CreateCommand())
-                {
-                    commandId.Transaction = transaction;
-                    commandId.CommandText = "SELECT last_insert_rowid();";
-                    id = Convert.ToInt64(commandId.ExecuteScalar());
-                }
-            }
-
-            if (id <= 0)
-            {
-                throw new DotNetWorkQueueException(
-                    "Failed to insert record - the ID of the new record returned by SQLite was 0");
-            }
-
-            using (var commandMeta = SendMessage.CreateMetaDataRecord(commandSend.MessageData.GetDelay(), expiration,
-                connection, commandSend.MessageToSend, commandSend.MessageData, _tableNameHelper, _headers,
-                _options.Value, _getTime))
-            {
-                commandMeta.Transaction = transaction;
-                var param = commandMeta.CreateParameter();
-                param.ParameterName = "@QueueID";
-                param.DbType = DbType.Int64;
-                param.Value = id;
-                commandMeta.Parameters.Add(param);
-                commandMeta.ExecuteNonQuery();
-            }
-
-            if (_options.Value.EnableStatusTable)
-            {
-                using (var commandStatus = CreateStatusRecord(connection, commandSend.MessageToSend, commandSend.MessageData))
-                {
-                    commandStatus.Transaction = transaction;
-                    var param = commandStatus.CreateParameter();
-                    param.ParameterName = "@QueueID";
-                    param.DbType = DbType.Int64;
-                    param.Value = id;
-                    commandStatus.Parameters.Add(param);
-                    commandStatus.ExecuteNonQuery();
-                }
-            }
-
-            if (!string.IsNullOrWhiteSpace(jobName))
-            {
-                _sendJobStatus.Handle(new SetJobLastKnownEventCommand<IDbConnection, IDbTransaction>(
-                    jobName, eventTime, scheduledTime, connection, transaction));
-            }
-
-            // Caller owns lifecycle: no Commit, Rollback, Close, or Dispose performed here.
-            return id;
         }
 
         /// <summary>
