+++ Source/DotNetWorkQueue.Transport.SQLite/Decorator/RetryCommandHandlerOutputDecoratorAsync.cs
@@ -18,7 +18,6 @@
 // ---------------------------------------------------------------------
 
 using System;
-using DotNetWorkQueue.Transport.RelationalDatabase;
 using DotNetWorkQueue.Transport.Shared;
 using DotNetWorkQueue.Transport.SQLite.Basic;
 using DotNetWorkQueue.Validation;
@@ -52,10 +51,6 @@ namespace DotNetWorkQueue.Transport.SQLite.Decorator
         public async Task<TOutput> HandleAsync(TCommand command)
         {
             Guard.NotNull(() => command, command);
-
-            if (command is IRetrySkippable skippable && skippable.SkipRetry)
-                return await _decorated.HandleAsync(command).ConfigureAwait(false);
-
             ResiliencePipeline pipeline = null;
             try
             {
